# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /events:
    get:
      operationId: EventsController_findAll
      parameters:
        - name: limit
          required: false
          in: query
          description: Number of items to return
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: offset
          required: false
          in: query
          description: Number of items to skip
          schema:
            minimum: 0
            default: 0
            type: number
        - name: search
          required: false
          in: query
          description: Search by event title
          schema:
            type: string
        - name: eventType
          required: false
          in: query
          description: Filter by event type
          schema:
            type: string
            enum:
              - INDUCTION
              - WORKSHOP
              - COMPETITION
              - SEMINAR
              - HACKATHON
              - MEETUP
              - OTHER
        - name: hasTeam
          required: false
          in: query
          description: Filter team-based events
          schema:
            type: boolean
        - name: upcoming
          required: false
          in: query
          description: Filter upcoming events only
          schema:
            type: boolean
        - name: past
          required: false
          in: query
          description: Filter past events only
          schema:
            type: boolean
      responses:
        '200':
          description: ''
      summary: List all events with filters and pagination
      tags:
        - Events
    post:
      operationId: EventsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventDto'
      responses:
        '201':
          description: ''
      security:
        - team-auth: []
      summary: Create a new event (Admin only)
      tags:
        - Events
  /events/{id}:
    get:
      operationId: EventsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get event by ID with schedules
      tags:
        - Events
    patch:
      operationId: EventsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventDto'
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Update event (Admin only)
      tags:
        - Events
    delete:
      operationId: EventsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Delete event (Admin only)
      tags:
        - Events
  /events/{id}/schedules:
    post:
      operationId: EventsController_addSchedule
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleDto'
      responses:
        '201':
          description: ''
      security:
        - team-auth: []
      summary: Add schedule to event (Admin only)
      tags:
        - Events
  /events/{eventId}/schedules/{scheduleId}:
    patch:
      operationId: EventsController_updateSchedule
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
        - name: scheduleId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScheduleDto'
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Update event schedule (Admin only)
      tags:
        - Events
    delete:
      operationId: EventsController_removeSchedule
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
        - name: scheduleId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Delete event schedule (Admin only)
      tags:
        - Events
  /events/{id}/register:
    post:
      operationId: EventsController_register
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      security:
        - user-auth: []
      summary: Register for an event as solo participant
      tags:
        - Events
  /events/{id}/unregister:
    delete:
      operationId: EventsController_unregister
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - user-auth: []
      summary: Unregister from an event (Authenticated participant)
      tags:
        - Events
  /events/{id}/participants:
    get:
      operationId: EventsController_getParticipants
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: Number of items to return
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: offset
          required: false
          in: query
          description: Number of items to skip
          schema:
            minimum: 0
            default: 0
            type: number
      responses:
        '200':
          description: ''
      summary: List event participants
      tags:
        - Events
  /participants/me:
    get:
      operationId: ParticipantsController_getMe
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - user-auth: []
      summary: Get current participant profile
      tags:
        - Participants
    post:
      operationId: ParticipantsController_createProfile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateParticipantDto'
      responses:
        '201':
          description: ''
      security:
        - user-auth: []
      summary: Create participant profile
      tags:
        - Participants
    patch:
      operationId: ParticipantsController_updateProfile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateParticipantDto'
      responses:
        '200':
          description: ''
      security:
        - user-auth: []
      summary: Update participant profile
      tags:
        - Participants
  /participants/me/events:
    get:
      operationId: ParticipantsController_getMyEvents
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - user-auth: []
      summary: Get events registered by current participant
      tags:
        - Participants
  /participants/{id}:
    get:
      operationId: ParticipantsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get participant public profile by ID
      tags:
        - Participants
...
  /teams/{id}/members/{participantId}:
    delete:
      operationId: TeamsController_removeMember
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: participantId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - user-auth: []
      summary: Remove member from team (Team leader only)
      tags:
        - Teams
  /members:
    get:
      operationId: MembersController_findAll
      parameters:
        - name: limit
          required: false
          in: query
          description: Number of items to return
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: offset
          required: false
          in: query
          description: Number of items to skip
          schema:
            minimum: 0
            default: 0
            type: number
        - name: search
          required: false
          in: query
          description: Search by name or username
          schema:
            type: string
        - name: position
          required: false
          in: query
          description: Filter by position title
          schema:
            type: string
        - name: collegeId
          required: false
          in: query
          description: Filter by college ID
          schema:
            type: string
        - name: departmentId
          required: false
          in: query
          description: Filter by department ID
          schema:
            type: string
        - name: graduationYear
          required: false
          in: query
          description: Filter by graduation year
          schema:
            type: number
      responses:
        '200':
          description: ''
      summary: List all club members with filters
      tags:
        - Members
  /members/{id}:
    get:
      operationId: MembersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get member by ID with positions
      tags:
        - Members
    patch:
      operationId: MembersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberDto'
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Update member profile (Self or Admin)
      tags:
        - Members
    delete:
      operationId: MembersController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Remove member (Admin only)
      tags:
        - Members
  /members/invite:
    post:
      description: >-
        Sends a Clerk invitation to the email. Member profile is created with pending status. The invitation is accepted
        via Clerk, which triggers a webhook to activate the member.
      operationId: MembersController_invite
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteMemberDto'
      responses:
        '201':
          description: ''
      security:
        - team-auth: []
      summary: Invite a new member (Admin only)
      tags:
        - Members
...
  /projects/{projectId}/members/{memberId}:
    delete:
      operationId: ProjectsController_removeMember
      parameters:
        - name: projectId
          required: true
          in: path
          schema:
            type: string
        - name: memberId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Remove member from project (Project member only)
      tags:
        - Projects
...
  /certificates/event/{eventId}/participant/{collegeIdNo}:
    get:
      operationId: CertificatesController_getCertificate
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
        - name: collegeIdNo
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      summary: Download certificate PDF for participant (rate limited)
      tags:
        - Certificates
...
info:
  title: The Robotics Club API
  description: API for the college robotics club management platform
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Local Development
  - url: https://api.roboticsclub.example.com
    description: Production
components:
  securitySchemes:
    user-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Clerk JWT token for participants (external users)
    team-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Clerk JWT token for club members/admins
  schemas:
...
    CreateEventDto:
      type: object
      properties:
        title:
          type: string
          description: Event title
          example: Robotics Workshop 2024
        description:
          type: string
          description: Event description
        eventType:
          type: string
          description: Event type
          enum:
            - INDUCTION
            - WORKSHOP
            - COMPETITION
            - SEMINAR
            - HACKATHON
            - MEETUP
            - OTHER
        hasTeam:
          type: boolean
          description: Whether this is a team-based event
          default: false
        maxTeamMembers:
          type: number
          description: Maximum team members (required if hasTeam is true)
          minimum: 2
        maxParticipants:
          type: number
          description: Maximum participants allowed
          minimum: 1
        registrationDeadline:
          type: string
          description: Registration deadline (ISO 8601 date)
          example: '2024-12-31T23:59:59Z'
      required:
        - title
        - description
        - eventType
        - registrationDeadline
    UpdateEventDto:
      type: object
      properties:
        title:
          type: string
          description: Event title
        description:
          type: string
          description: Event description
        eventType:
          type: string
          description: Event type
          enum:
            - INDUCTION
            - WORKSHOP
            - COMPETITION
            - SEMINAR
            - HACKATHON
            - MEETUP
            - OTHER
        hasTeam:
          type: boolean
          description: Whether this is a team-based event
        maxTeamMembers:
          type: number
          description: Maximum team members
          minimum: 2
        maxParticipants:
          type: number
          description: Maximum participants
          minimum: 1
        registrationDeadline:
          type: string
          description: Registration deadline (ISO 8601 date)
    CreateScheduleDto:
      type: object
      properties:
        day:
          type: string
          description: Schedule date (ISO 8601)
          example: '2024-12-01'
        startTime:
          type: string
          description: Start time (HH:MM)
          example: '09:00'
        endTime:
          type: string
          description: End time (HH:MM)
          example: '17:00'
        location:
          type: string
          description: Location
          example: Main Auditorium
        notes:
          type: string
          description: Additional notes
      required:
        - day
        - startTime
        - endTime
        - location
    UpdateScheduleDto:
      type: object
      properties:
        day:
          type: string
          description: Schedule date (ISO 8601)
        startTime:
          type: string
          description: Start time (HH:MM)
        endTime:
          type: string
          description: End time (HH:MM)
        location:
          type: string
          description: Location
        notes:
          type: string
          description: Additional notes
    CreateParticipantDto:
      type: object
      properties:
        name:
          type: string
          description: Full name
          example: John Doe
        email:
          type: string
          description: Email address
          example: john@example.com
        username:
          type: string
          description: Username (lowercase, numbers, underscores only)
          example: john_doe
        imageUrl:
          type: string
          description: Profile image URL
        phone:
          type: string
          description: Phone number
        collegeId:
          type: string
          description: College ID
        departmentId:
          type: string
          description: Department ID
        collegeIdNo:
          type: number
          description: College ID number (roll number)
          example: 12345
        year:
          type: string
          description: Academic year
          example: 3rd Year
      required:
        - name
        - email
        - username
        - imageUrl
        - collegeId
        - departmentId
        - collegeIdNo
    UpdateParticipantDto:
      type: object
      properties:
        name:
          type: string
          description: Full name
        imageUrl:
          type: string
          description: Profile image URL
        phone:
          type: string
          description: Phone number
        year:
          type: string
          description: Academic year
...
    InviteMemberDto:
      type: object
      properties:
        name:
          type: string
          description: Full name
          example: Jane Smith
        email:
          type: string
          description: Email address
          example: jane@example.com
        username:
          type: string
          description: Username (lowercase, numbers, underscores only)
          example: jane_smith
        collegeId:
          type: string
          description: College ID
        departmentId:
          type: string
          description: Department ID
        collegeIdNo:
          type: number
          description: College ID number (roll number)
          example: 12345
        graduationYear:
          type: number
          description: Expected graduation year
          example: 2025
          minimum: 2000
          maximum: 2100
        imageUrl:
          type: string
          description: Profile image URL
      required:
        - name
        - email
        - username
        - collegeId
        - departmentId
        - collegeIdNo
        - imageUrl
    UpdateMemberDto:
      type: object
      properties:
        name:
          type: string
          description: Full name
        imageUrl:
          type: string
          description: Profile image URL
        phone:
          type: string
          description: Phone number
        bio:
          type: string
          description: Bio
        graduationYear:
          type: number
          description: Graduation year
          minimum: 2000
          maximum: 2100
        instagram:
          type: string
          description: Instagram handle
        linkedin:
          type: string
          description: LinkedIn URL
        github:
          type: string
          description: GitHub username
...